Backend/requirements.txt
python-dotenv
pymongo[srv]
fastapi[standard]
motor-types
motor
uvicorn